Run app with non-root user
Running applications in Docker as root user opens various attack
vectors, as Docker's "isolation" isn't meant to be the only safe guard
against adversaries. It should be considered a "last resort". It's best
practice to use a non-root user wherever possible.

This commit changes back to Alpine as a base, as it provides the adduser
tool via busybox. It increases the image size by a few MiB (2-3,
currently). One could optimize the image size by writing a custom /etc/passwd,
but this seems quite unnecessary, especially given that Alpine is the
base of so many other images.